% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docore-package.R
\docType{package}
\name{docore}
\alias{docore}
\title{Utility Functions for Scientific Data Analysis}
\description{
Basic routines used in scientific coding, such as timing routines, vector/array handing functions and I/O support routines.
}
\author{
Danail Obreschkow
}
